Updated tests to use gulp 4.0 syntaxes
Previous gulpfile changed to gulpfile-3.8.10
