+++ test/flags-continue.js
@@ -9,7 +9,7 @@ var child = require('child_process');
 lab.experiment('flag: --continue', function() {
 
   lab.test('continues execution when flag is set', function(done) {
-    child.exec('node ' + __dirname + '/../bin/gulp.js test4 --continue --gulpfile "./test/fixtures/gulpfile-4-0.js"', function(err, stdout, stderr) {
+    child.exec('node ' + __dirname + '/../bin/gulp.js test4 --continue --cwd ./test/fixtures', function(err, stdout, stderr) {
       code.expect(stdout).to.contain('Starting \'errorFunction\'');
       code.expect(stderr).to.contain('\'errorFunction\' errored after');
       stdout = stdout.replace(/\\/g, '/').split('\n');
@@ -19,8 +19,8 @@ lab.experiment('flag: --continue', function() {
     });
   });
 
-  lab.test('continues execution when flag is set', function(done) {
-    child.exec('node ' + __dirname + '/../bin/gulp.js test4 --gulpfile "./test/fixtures/gulpfile-4-0.js"', function(err, stdout, stderr) {
+  lab.test('stops execution when flag is not set', function(done) {
+    child.exec('node ' + __dirname + '/../bin/gulp.js test4 --cwd ./test/fixtures', function(err, stdout, stderr) {
       code.expect(stdout).to.contain('Starting \'errorFunction\'');
       code.expect(stderr).to.contain('\'errorFunction\' errored after');
       code.expect(stdout[4]).to.not.contain('Starting \'anon\'');
