+++ test/flags-gulpfile.js
@@ -12,7 +12,7 @@ lab.experiment('flag: --gulpfile', function() {
     child.exec('node ' + __dirname + '/../bin/gulp.js --gulpfile "./test/fixtures/gulpfile-2.js"', function(err, stdout) {
       stdout = stdout.replace(/\\/g, '/').split('\n');
       code.expect(stdout[1]).to.contain('test/fixtures/gulpfile-2.js');
-      code.expect(stdout[3]).to.contain('Finished \'default\'');
+      code.expect(stdout[5]).to.contain('Finished \'default\'');
       done(err);
     });
   });
