+++ test/flags-tasks-json.js
@@ -10,7 +10,7 @@ var output = require('./expected/flags-tasks-json.json')
 lab.experiment('flag: --tasks-json', function() {
 
   lab.test('prints the task list with no args', function(done) {
-    child.exec('node ' + __dirname + '/../bin/gulp.js --tasks-json --gulpfile "./test/fixtures/gulpfile-4-0.js" ', function(err, stdout) {
+    child.exec('node ' + __dirname + '/../bin/gulp.js --tasks-json --gulpfile "./test/fixtures/gulpfile.js" ', function(err, stdout) {
       stdout = stdout.replace(/\\/g, '/').split('\n');
       var parsedJson = JSON.parse(stdout[1]);
       code.expect(parsedJson).to.deep.equal(output);
@@ -24,7 +24,7 @@ lab.experiment('flag: --tasks-json', function() {
         return done(err);
       }
 
-      child.exec('node ' + __dirname + '/../bin/gulp.js --tasks-json "../output/tasks.json" --gulpfile "./test/fixtures/gulpfile-4-0.js" ', function(err, stdout) {
+      child.exec('node ' + __dirname + '/../bin/gulp.js --tasks-json "../output/tasks.json" --gulpfile "./test/fixtures/gulpfile.js" ', function(err, stdout) {
         var file = fs.readFileSync(__dirname + '/output/tasks.json', 'utf8')
         var parsedJson = JSON.parse(file);
         code.expect(parsedJson).to.deep.equal(output);
