+++ test/flags-tasks.js
@@ -2,8 +2,10 @@
 
 var lab = exports.lab = require('lab').script();
 var code = require('code');
-
+var fs = require('fs');
 var child = require('child_process');
+
+var output = fs.readFileSync(__dirname + '/expected/flags-tasks.txt', 'utf8').replace(/\r\n/g, '\n');
 
 lab.experiment('flag: --tasks', function() {
 
@@ -11,14 +13,10 @@ lab.experiment('flag: --tasks', function() {
     child.exec('node ' + __dirname + '/../bin/gulp.js --tasks --cwd ./test/fixtures', function(err, stdout) {
       code.expect(stdout).to.contain('Tasks for');
       stdout = stdout.replace(/\\/g, '/').split('Tasks for')[1].split('\n');
-      code.expect(stdout[0]).to.contain('/gulp-cli/test');
-      code.expect(stdout[1]).to.contain('├── test1');
-      code.expect(stdout[2]).to.contain('├─┬ test2');
-      code.expect(stdout[3]).to.contain('│ └── test1');
-      code.expect(stdout[4]).to.contain('├── test3    description');
-      code.expect(stdout[5]).to.contain('└─┬ default');
-      code.expect(stdout[6]).to.contain('  ├── test1');
-      code.expect(stdout[7]).to.contain('  └── test3');
+      var outputArray = output.split('\n');
+      for (var i = 0; i < stdout.length; i++) {
+        code.expect(stdout[i]).to.contain(outputArray[i]);
+      }
       done(err);
     });
   });
